add ajax post

+++ src/main/resources/templates/ajax/feeds.js
@@ -9,24 +9,16 @@ function findUserById() {
 }
 
 function deletePost(id) {
-    $.ajax({
-        type: "delete",
-        url: `http://localhost:8080/api/post/${id}`,
-        success: function () {
-            showAllPost();
-        }
-    });
-}
+    if (confirm('Are you sure you want to delete ?') === true) {
+        $.ajax({
+            type: "delete",
+            url: `http://localhost:8080/api/post/${id}`,
+            success: function () {
+                showAllPost();
+            }
+        });
+    }
 
-function countlikePost(id) {
-    $.ajax({
-        type: "GET",
-        url: `http://localhost:8080/api/like-post/${id}`,
-        success: function (data) {
-            console.log(data)
-            document.getElementById("likeCount" + id ).innerHTML = data;
-        }
-    });
 }
 
 function showAllPost() {
@@ -46,7 +38,9 @@ function showAllPost() {
     })
 }
 
+
 function displayPost(post) {
+    console.log(post)
     let content = "";
     content += `<div class="post">
                                 <div class="post-heading">
@@ -61,7 +55,7 @@ function displayPost(post) {
                                         <span class="icon-more uil-ellipsis-h"></span>
                                         <div class="mt-0 p-2" uk-dropdown="pos: top-right;mode:hover ">
                                             <ul class="uk-nav uk-dropdown-nav">
-                                                <li><a onclick="editPost()"> <i class="uil-edit-alt mr-1"></i> Edit Post </a></li>
+                                                <li><a onclick="editPost(${post.id})"> <i class="uil-edit-alt mr-1"></i> Edit Post </a></li>
                                            
                                                 <li><a onclick="deletePost(${post.id})" class="text-danger"> <i class="uil-trash-alt mr-1"></i>
                                                     Delete </a></li>
@@ -71,40 +65,41 @@ function displayPost(post) {
                                 </div>
                                 <div class="post-description">
                                     <div class="fullsizeimg">
-                                    <p>${post.content}</p>`;
+                                    <p style="padding-left: 20px">${post.content}</p>`;
     if (post.imgUrl != null) {
         content += `<img src="${post.imgUrl}" alt="">`;
     }
     content += `</div>
                                 </div>
                                 <div class="post-state">
-                                    <div class="post-state-btns"> <i class="uil-thumbs-up"></i> <span id="likeCount"> </span> <span> Liked </span>
+                                    <div class="post-state-btns"> 
+                                    <i class="uil-thumbs-up"></i> 
+                                    <span id="likeCount${post.id}" style="padding-right: 5px"></span>
+                                    <a onclick="likePost(${post.id},1)" class="view-more-comment"> Like </a>
                                     </div>
                                 </div>
 
                                 <!-- post comments -->
                                 <div class="post-comments">
-                                    <a onclick="showComments(${post.id})" class="view-more-comment"> View Comments</a>
-                                    <div id="commentList"> 
-                                        
-                                    </div>
-                                    
+                                    <a onclick="showComments(${post.id})" class="view-more-comment">View Comments</a>
+                                    <div id="commentList${post.id}"></div>
                                     <div class="post-add-comment">
                                         <div class="post-add-comment-text-area">
-                                            <input type="text" placeholder="Write your comment...">
+                                            <input type="text" id="commentPost" placeholder="Write your comment...">
+                                            <button type="button" onclick="createCommentPost(${post.id},1)" class="button primary px-6"> Comment </button>
                                         </div>
 
                                     </div>
 
                                 </div>
-                            </div>`
+                            </div>
+<div id="commentList"></div>`
 
     countlikePost(post.id);
     return content;
 }
 
 function showComments(id) {
-    console.log(id)
     $.ajax({
         type: "GET",
         url: `http://localhost:8080/api/comment/${id}`,
@@ -112,16 +107,17 @@ function showComments(id) {
             let content = "";
             if (data !== undefined) {
                 for (let i = 0; i < data.length; i++) {
-                    content += displayContent(data[i]);
+                    content += displayComment(data[i]);
                 }
             }
-            document.getElementById("commentList").innerHTML = content;
+            console.log(data)
+            document.getElementById("commentList" + id).innerHTML = content;
         }
     })
 }
 
-function displayContent(comment) {
-   return `<div class="post-comments-single">
+function displayComment(comment) {
+    let content = `<div class="post-comments-single">
         <div class="post-comment-avatar">
             <img src="${comment.userInfo.avatarUrl}" alt="">
         </div>
@@ -130,13 +126,172 @@ function displayContent(comment) {
                 <h6> ${comment.userInfo.fullName} </h6>
                 <p> ${comment.content} </p>
             </div>
-            <div class="uk-text-small">
-                <a onclick="likeComment()" class="text-danger mr-1"> 
-                <i class="uil-thumbs-up"></i> like
-                </a>
-            </div>
+            <div class="uk-text-small"> 
+                                    <i class="uil-thumbs-up"></i> 
+                                    <span id="likeCountComment${comment.id}" style="padding-right: 5px"></span>
+                                    <a onclick="likeComment(${comment.id},1)" class="view-more-comment"> Like </a>
+                                    </div>
         </div>
     </div>`
+
+    countlikeComment(comment.id);
+    return content;
+}
+
+function countlikePost(id) {
+    $.ajax({
+        type: "GET",
+        url: `http://localhost:8080/api/like-post/${id}`,
+        success: function (data) {
+            // console.log(data)
+            document.getElementById("likeCount" + id).innerHTML = data;
+        }
+    });
+}
+
+function countlikeComment(id) {
+    $.ajax({
+        type: "GET",
+        url: `http://localhost:8080/api/like-comment/${id}`,
+        success: function (data) {
+            console.log(data)
+            document.getElementById("likeCountComment" + id).innerHTML = data;
+        }
+    });
+}
+
+function likeComment(comment_id, userinfo_id) {
+    $.ajax({
+        type: "post",
+        url: `http://localhost:8080/api/like-comment/${comment_id}/${userinfo_id}`,
+        success: function () {
+            showAllPost();
+        }
+    });
+}
+
+function likePost(post_id, userinfo_id) {
+    $.ajax({
+        type: "post",
+        url: `http://localhost:8080/api/like-post/${post_id}/${userinfo_id}`,
+        success: function () {
+            showAllPost();
+        }
+    });
+}
+
+function createPost(user_id) {
+    let data = new FormData();
+    let content = $('#contentPost').val();
+    let status = $('#statusPost').val();
+    let newPost = {
+        content: content,
+        status: status
+    };
+
+    data.append("file", $('#fileImage')[0].files[0]);
+    data.append("post", new Blob([JSON.stringify(newPost)], {
+        type: "application/json"
+    }))
+
+    //gọi ajax
+    $.ajax({
+        type: "POST",
+        data: data,
+        processData: false,
+        contentType: false,
+        //tên API
+        url: `http://localhost:8080/api/post/${user_id}`,
+        //xử lý khi thành công
+        success: function () {
+            showAllPost();
+            document.getElementById("formCreatePost").reset();
+        }
+
+    });
+    //chặn sự kiện mặc định của thẻ
+    event.preventDefault();
+
+}
+let index = 0;
+function editPost(id) {
+    $.ajax({
+        type: 'GET',
+        url: `http://localhost:8080/api/post/${id}`,
+        success: function (data) {
+            $('#contentPost').val(data.content);
+            $('#statusPost').val(data.status);
+            index = data.id;
+            document.getElementById("formSubmit").onclick = function () {
+                updatePost();
+            };
+        }
+    });
+}
+
+function updatePost() {
+    let data = new FormData();
+    let content = $('#contentPost').val();
+    let status = $('#statusPost').val();
+    let newPost = {
+        content: content,
+        status: status
+    };
+
+    data.append("file", $('#fileImage')[0].files[0]);
+    data.append("postUpdate", new Blob([JSON.stringify(newPost)], {
+        type: "application/json"
+    }))
+
+    //gọi ajax
+    $.ajax({
+        type: "put",
+        data: data,
+        processData: false,
+        contentType: false,
+        //tên API
+        url: `http://localhost:8080/api/post/${index}`,
+        //xử lý khi thành công
+        success: function () {
+            showAllPost();
+            document.getElementById("formCreatePost").reset();
+        }
+
+    });
+    //chặn sự kiện mặc định của thẻ
+    event.preventDefault();
+}
+
+function createCommentPost(post_id, userinfo_id) {
+    let content = $('#commentPost').val();
+    let newComment = {
+        content: content,
+        userInfo: {
+            id: userinfo_id
+        },
+        post: {
+            id: post_id
+        }
+    };
+
+    //gọi ajax
+    $.ajax({
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        },
+        type: "POST",
+        data: JSON.stringify(newComment),
+        //tên API
+        url: `http://localhost:8080/api/comment`,
+        //xử lý khi thành công
+        success: function () {
+            showAllPost();
+        }
+
+    });
+    //chặn sự kiện mặc định của thẻ
+    event.preventDefault();
 }
 
 window.onload = showAllPost;
